moving utils.go code to router.go
