+++ router/router.go
@@ -4,13 +4,15 @@ import (
 	"io"
 	"log"
 	"os"
+	"text/template"
+	"time"
 
 	"github.com/gin-contrib/sessions"
 	gormsessions "github.com/gin-contrib/sessions/gorm"
 	"github.com/gin-gonic/gin"
 	"github.com/nyudlts/go-medialog/config"
+	"github.com/nyudlts/go-medialog/controllers"
 	"github.com/nyudlts/go-medialog/database"
-	"github.com/nyudlts/go-medialog/utils"
 )
 
 func SetupRouter(env config.Environment, gormDebug bool, prod bool) (*gin.Engine, error) {
@@ -32,7 +34,7 @@ func SetupRouter(env config.Environment, gormDebug bool, prod bool) (*gin.Engine
 	r := gin.Default()
 
 	//add global funcs
-	utils.SetGlobalFuncs(r)
+	SetGlobalFuncs(r)
 
 	//configure the router
 	r.LoadHTMLGlob("templates/**/*.html")
@@ -62,4 +64,30 @@ func SetupRouter(env config.Environment, gormDebug bool, prod bool) (*gin.Engine
 	LoadRoutes(r)
 
 	return r, nil
+}
+
+func Add(a int, b int) int { return a + b }
+
+func Subtract(a int, b int) int { return a - b }
+
+func Multiply(a int, b int) int { return a * b }
+
+func MultAndAdd(page int, mult int, add int) int {
+	return Add(add, Multiply(page, mult))
+}
+
+func FormatAsDate(t time.Time) string { return t.Format("2006-01-02") }
+
+func SetGlobalFuncs(router *gin.Engine) {
+	router.SetFuncMap(template.FuncMap{
+		"formatAsDate":       FormatAsDate,
+		"add":                Add,
+		"subtract":           Subtract,
+		"multiply":           Multiply,
+		"getMediatype":       controllers.GetMediaType,
+		"getMediatypes":      controllers.GetMediatypes,
+		"multAndAdd":         MultAndAdd,
+		"storageLocations":   controllers.GetStorageLocations,
+		"getStorageLocation": controllers.GetStorageLocation,
+	})
 }
