+++ templates/entries/entries-show.html
@@ -140,5 +140,6 @@
         </table>
     </div>
 </div>
+<br>
 {{ template "entries-toolbar.html" .}}
 {{ template "footer.html" .}}
